修正在 Linux 环境下的兼容问题, 使之能正常工作
