+++ CMakeLists.txt
@@ -237,7 +237,7 @@ set( Boost_USE_STATIC_RUNTIME OFF ) # 不使用 runtime 静态库
 set( BOOST_ROOT 3rdparty/boost ) # 从指定的位置搜索 Boost 库
 
 # 寻找名为 Boost 的库, 要求携带指定的多个组件, 失败则报错
-find_package(Boost COMPONENTS regex locale filesystem date_time REQUIRED)
+find_package(Boost COMPONENTS regex locale filesystem date_time serialization REQUIRED)
 
 # 如果成功找到 Boost 库, 那么进行相关的设置操作
 if( Boost_FOUND )
