+++ src/map/map-server.vcxproj.filters
@@ -173,9 +173,6 @@
     <ClInclude Include="aura.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="serialize.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="achievement.cpp">
