+++ src/map/skill.hpp
@@ -16,7 +16,7 @@
 #include "map.hpp" // struct block_list
 
 #ifdef Pandas_YamlBlastCache_Serialize
-#include "serialize.hpp"
+#include "../common/serialize.hpp"
 #endif // Pandas_YamlBlastCache_Serialize
 
 enum damage_lv : uint8;
@@ -310,9 +310,9 @@ private:
 	friend class boost::serialization::access;
 
 	template <typename Archive>
-	friend void boost::serialization::serialize(
-		Archive& ar, SkillDatabase& t, const unsigned int version
-	);
+	void serialize(Archive& ar, const unsigned int version) {
+		ar& boost::serialization::base_object<TypesafeCachedYamlDatabase<uint16, s_skill_db>>(*this);
+	}
 #endif // Pandas_YamlBlastCache_SkillDatabase
 public:
 	SkillDatabase() : TypesafeCachedYamlDatabase("SKILL_DB", 1) {
@@ -2380,5 +2380,107 @@ int skill_get_time3(struct map_data *mapdata, uint16 skill_id, uint16 skill_lv);
 #define SKILL_CHK_MERC(skill_id)  ( (skill_id) >= MC_SKILLBASE && (skill_id) < MC_SKILLBASE+MAX_MERCSKILL )
 #define SKILL_CHK_ELEM(skill_id)  ( (skill_id) >= EL_SKILLBASE && (skill_id) < EL_SKILLBASE+MAX_ELEMENTALSKILL )
 #define SKILL_CHK_GUILD(skill_id) ( (skill_id) >= GD_SKILLBASE && (skill_id) < GD_SKILLBASE+MAX_GUILDSKILL )
+
+#ifdef Pandas_YamlBlastCache_SkillDatabase
+namespace boost {
+	namespace serialization {
+		// ======================================================================
+		// struct s_skill_db
+		// ======================================================================
+
+		template <typename Archive>
+		void serialize(Archive& ar, struct s_skill_db& t, const unsigned int version)
+		{
+			ar& t.nameid;
+			ar& t.name;
+			ar& t.desc;
+			ar& t.range;
+			ar& t.hit;
+			ar& t.inf;
+			ar& t.element;
+			ar& t.nk;
+			ar& t.splash;
+			ar& t.max;
+			ar& t.num;
+			ar& t.castcancel;
+			ar& t.cast_def_rate;
+			ar& t.skill_type;
+			ar& t.blewcount;
+			ar& t.inf2;
+			ar& t.maxcount;
+
+			ar& t.castnodex;
+			ar& t.delaynodex;
+
+			//ar& t.nocast;					// SkillDatabase 默认不会为其赋值, 暂时无需处理
+
+			ar& t.unit_id;
+			ar& t.unit_id2;
+			ar& t.unit_layout_type;
+			ar& t.unit_range;
+			ar& t.unit_interval;
+			ar& t.unit_target;
+			ar& t.unit_flag;
+
+			ar& t.cast;
+			ar& t.delay;
+			ar& t.walkdelay;
+			ar& t.upkeep_time;
+			ar& t.upkeep_time2;
+			ar& t.cooldown;
+#ifdef RENEWAL_CAST
+			ar& t.fixed_cast;
+#endif // RENEWAL_CAST
+
+			ar& t.require;
+
+			ar& t.unit_nonearnpc_range;
+			ar& t.unit_nonearnpc_type;
+
+			//ar& t.damage;					// SkillDatabase 默认不会为其赋值, 暂时无需处理
+			ar& t.copyable;
+
+			//ar& t.abra_probability;		// SkillDatabase 默认不会为其赋值, 暂时无需处理
+			//ar& t.reading_spellbook;		// SkillDatabase 默认不会为其赋值, 暂时无需处理
+			//ar& t.improvisedsong_rate;	// SkillDatabase 默认不会为其赋值, 暂时无需处理
+		}
+
+		// ======================================================================
+		// struct s_skill_copyable
+		// ======================================================================
+
+		template <typename Archive>
+		void serialize(Archive& ar, struct s_skill_copyable& t, const unsigned int version)
+		{
+			ar& t.option;
+			ar& t.req_opt;
+		}
+
+		// ======================================================================
+		// struct s_skill_require
+		// ======================================================================
+
+		template <typename Archive>
+		void serialize(Archive& ar, struct s_skill_require& t, const unsigned int version)
+		{
+			ar& t.hp;
+			ar& t.mhp;
+			ar& t.sp;
+			ar& t.hp_rate;
+			ar& t.sp_rate;
+			ar& t.zeny;
+			ar& t.weapon;
+			ar& t.ammo;
+			ar& t.ammo_qty;
+			ar& t.state;
+			ar& t.spiritball;
+			ar& t.itemid;
+			ar& t.amount;
+			ar& t.eqItem;
+			ar& t.status;
+		}
+	} // namespace serialization
+} // namespace boost
+#endif // Pandas_YamlBlastCache_SkillDatabase
 
 #endif /* SKILL_HPP */
